requirements.txt
vosk==0.3.45
PyAudio==0.2.13
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dateutil==2.8.2
dateparser==1.1.8
python-dotenv==1.0.0
pytz==2023.3